databaseUpdate/requirements.txt
dash==2.14.2
dash_bootstrap_components==1.5.0
geopy==2.4.1
matplotlib==3.8.2
numpy==1.24.3
openpyxl==3.0.10
pandas==2.2.0
Pillow==10.2.0
plotly==5.18.0
psycopg2==2.9.3
pyproj==3.6.1
requests==2.31.0
seaborn==0.13.2
gunicorn
dash-tools
Flask-SQLAlchemy
pyarrow
scikit-learn
